Route Permissions

+++ routes/api.php
@@ -12,27 +12,31 @@ Route::post('/players', [UserController::class, 'register']);
 Route::post('/login', [UserController::class, 'login'])->name('login');
 
 Route::middleware('auth:api')->group(function () {
-
+    
     Route::post('/logout', [UserController::class, 'logout'])->name('logout');
 
-    Route::put('/players/{id}', [UserController::class, 'update'])->name('users.update');
+    Route::put('/players/{id}', [UserController::class, 'update'])->name('users.update')->middleware('role:admin');
+    Route::put('/players/{id}', [UserController::class, 'update'])->name('users.update')->middleware('role:player');
 
     //Tirada Daus
-    Route::post('/players/{id}/games', [GameController::class, 'throwDice'])->name('games.throwDice');
+    Route::post('/players/{id}/games', [GameController::class, 'throwDice'])->name('games.throwDice')->middleware('role:admin');
+    Route::post('/players/{id}/games', [GameController::class, 'throwDice'])->name('games.throwDice')->middleware('role:player');
 
     //Elimina les tirades del jugador/a
-    Route::delete('/players/{id}/games', [GameController::class, 'delete'])->name('games.deletePlayerGames');
+    Route::delete('/players/{id}/games', [GameController::class, 'delete'])->name('games.deletePlayerGames')->middleware('role:admin');
+    Route::delete('/players/{id}/games', [GameController::class, 'delete'])->name('games.deletePlayerGames')->middleware('role:player');
 
     // Llistat de tots els jugadors/es del sistema amb el seu percentatge mitjà d’èxits
-    Route::get('/players', [UserController::class, 'index'])->name('users.index');
+    Route::get('/players', [UserController::class, 'index'])->name('users.index')->middleware('role:admin');
 
     //Llistat jugades X jugador ID
-    Route::get('/players/{id}/games', [GameController::class, 'getPlayerGames'])->name('games.getPlayerGames');
+    Route::get('/players/{id}/games', [GameController::class, 'getPlayerGames'])->name('games.getPlayerGames')->middleware('role:admin');
+    Route::get('/players/{id}/games', [GameController::class, 'getPlayerGames'])->name('games.getPlayerGames')->middleware('role:player');
 
     //Rànquing mitjà de tots els jugadors/es del sistema.
-    Route::get('/players/ranking', [UserController::class, 'getPlayersRanking'])->name('players.ranking');
+    Route::get('/players/ranking', [UserController::class, 'getPlayersRanking'])->name('players.ranking')->middleware('role:admin');
 
-    Route::get('/players/ranking/loser', [UserController::class, 'getWorstPlayer'])->name('players.ranking/loser');
+    Route::get('/players/ranking/loser', [UserController::class, 'getWorstPlayer'])->name('players.ranking/loser')->middleware('role:admin');
 
-    Route::get('/players/ranking/winner', [UserController::class, 'getBestPlayer'])->name('players.ranking/winner');
+    Route::get('/players/ranking/winner', [UserController::class, 'getBestPlayer'])->name('players.ranking/winner')->middleware('role:admin');
 });
